+++ interviewer-preview/renderer/assets/__vite-browser-external-BR0ckfGP.js
@@ -1,7 +1,0 @@
-import { i as __commonJSMin } from "./index-BhFL_T49.js";
-//#region __vite-browser-external
-var require___vite_browser_external = /* @__PURE__ */ __commonJSMin(((exports, module) => {
-	module.exports = {};
-}));
-//#endregion
-export default require___vite_browser_external();