+++ src/components/sections/CategoricalBinPrompts/PromptFields.jsx
@@ -194,6 +194,7 @@ const selectOptionProps = PropTypes.shape({
   label: PropTypes.string.isRequired,
   value: PropTypes.oneOfType([
     PropTypes.string,
+    PropTypes.number,
     PropTypes.array,
     PropTypes.bool,
   ]),